`reparentutil`: move `validateAllTabletsManaged` to `util.go`
Both reparenters call it now, so it belongs with the other shared tabletMap
helpers rather than in the ERS file. Test moves with it.
