+++ go/vt/vtctl/reparentutil/emergency_reparenter.go
@@ -186,27 +186,6 @@ func validateEmergencyReparentOptions(opts EmergencyReparentOptions) error {
 }
 
 // reparentShardLocked performs Emergency Reparent Shard operation assuming that the shard is already locked
-// validateAllTabletsManaged errors when any tablet in the shard reports TabletMode_UNMANAGED.
-// Vitess cannot stop replication on such a tablet or revoke its writes, so it cannot establish the
-// invariant a reparent reports, and a promotion could leave the shard with two writers. Tablets
-// written before TabletMode existed report MANAGED, so a shard mid-upgrade never trips this.
-func validateAllTabletsManaged(tabletMap map[string]*topo.TabletInfo) error {
-	var unmanaged []string
-	for alias, tabletInfo := range tabletMap {
-		if tabletInfo.GetMode() == topodatapb.TabletMode_UNMANAGED {
-			unmanaged = append(unmanaged, alias)
-		}
-	}
-	if len(unmanaged) == 0 {
-		return nil
-	}
-	// Map iteration order is random, so sort to keep the error stable.
-	slices.Sort(unmanaged)
-	return vterrors.Errorf(vtrpc.Code_FAILED_PRECONDITION,
-		"shard has unmanaged tablets %v that Vitess cannot revoke writes from, so it cannot be reparented safely; "+
-			"unmanaged tablets belong in a keyspace of their own", unmanaged)
-}
-
 func (erp *EmergencyReparenter) reparentShardLocked(ctx context.Context, ev *events.Reparent, keyspace, shard string, opts EmergencyReparentOptions) (err error) {
 	// log the starting of the operation and increment the counter
 	erp.logger.Infof("will initiate emergency reparent shard in keyspace - %s, shard - %s", keyspace, shard)
